+++ services/allowed-users-finder.js
@@ -31,6 +31,9 @@ function AllowedUsersFinder(renderingId, opts) {
               logger.error('Cannot retrieve the project you\'re trying to ' +
                 'unlock. Can you check that you properly copied the Forest ' +
                 'envSecret in the forest_liana initializer?');
+            } else if (result.status === 422) {
+              logger.error('Cannot retrieve the project you\'re trying to ' +
+                'unlock. The envSecret and renderingId seems to be missing or inconsistent.');
             } else {
               logger.error('Cannot retrieve any users for the project ' +
                 'you\'re trying to unlock. An error occured in Forest API.',