+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change Log
 
 ## [Unreleased]
-
 ## RELEASE 1.3.3 - 2017-10-03
+### Fixed
+- Intercom - Make the conversation details accessible.
 
 ## RELEASE 1.3.2 - 2017-10-02
 ### Fixed