+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Changed
 - Smart Relationships - Add a warning if a Smart Collection does not define the "idField" attribute necessary for Smart Relationships. 
+- Smart Fields - Prevent the Smart Fields computation errors to generate a crash and handle it letting the value empty.
 
 ## RELEASE 1.3.6 - 2017-10-11
 ### Changed